Refactor and move qr.bids_inject to bids.inject package
