+++ docs/source/api/index.rst
@@ -10,10 +10,10 @@ API Reference
    :toctree: generated
 
    audio.audiocodes
+   bids.inject
    bids.inject_sidecar
    bids.media
    capture.nosignal
-   qr.bids_inject
    qr.disp_mon
    qr.psychopy
    qr.qr_parse
